Rename modules and fix ReportHandler.
